Cleanup result presentation model. Extract to results.ts
