+++ src/main/resources/admin/tools/part-finder/part-finder.ts
@@ -8,7 +8,7 @@ import { hasRole as hasAuthRole } from "/lib/xp/auth";
 import { list as listApps } from "/lib/xp/app";
 import { getComponent, type ComponentDescriptorType } from "/lib/xp/schema";
 import { find, notNullOrUndefined, runAsAdmin, startsWith, stringAfterLast, unique } from "/lib/part-finder/utils";
-import type { ComponentList, MultiUsageInstance, ContentUsage } from "./part-finder.freemarker";
+import type { ComponentList } from "./part-finder.freemarker";
 import type { ComponentViewParams } from "/admin/views/component-view/component-view.freemarker";
 import type { Header, Link } from "/admin/views/header/header.freemarker";
 import { createEditorFunc, EditorResult } from "/admin/tools/part-finder/editor";
@@ -18,6 +18,7 @@ import {
   getComponentUsagesInRepo,
   listComponentsInApplication,
 } from "/admin/tools/part-finder/listing";
+import { buildContentResult } from "/admin/tools/part-finder/results";
 
 export type PartFinderQueryParams = {
   key: string;
@@ -201,52 +202,6 @@ const parseComponentPathsPerId = (targetIds) => {
   return componentPathsPerId;
 };
 
-const insertAndGetSummaryContent = (contents: ContentUsage[], result: EditorResult): ContentUsage => {
-  let currentContent: ContentUsage = contents.filter((content) => content.id === result.id)[0];
-
-  if (!currentContent) {
-    currentContent = {
-      id: result.id,
-      url: result.url,
-      displayName: result.displayName,
-      path: result.path,
-      multiUsage: [],
-    };
-    contents.push(currentContent);
-  }
-
-  return currentContent;
-};
-
-const setMultiUsage = (currentContent: ContentUsage, result: EditorResult) => {
-  if ("string" === typeof result.componentPath) {
-    const usage: MultiUsageInstance = { path: result.componentPath };
-    if (result.error) {
-      usage.error = result.error;
-    }
-
-    currentContent.multiUsage.push(usage);
-    setHasMultiUsage(currentContent, true);
-
-  } else if (Array.isArray(result.componentPath)) {
-    const usages: MultiUsageInstance[] = result.componentPath.map((usage) => {
-      const res: MultiUsageInstance = { path: usage };
-      if (result.error) {
-        res.error = result.error;
-      }
-      return res;
-    });
-
-    currentContent.multiUsage.push(...usages);
-    setHasMultiUsage(currentContent, true);
-  } else if (result.componentPath === null) {
-    if (result.error) {
-      currentContent.error = result.error;
-    }
-    setHasMultiUsage(currentContent, false);
-  }
-};
-
 export function post(req: XP.Request): XP.Response {
   if (!hasAuthRole("system.admin")) {
     return {
@@ -371,21 +326,7 @@ export function post(req: XP.Request): XP.Response {
   const appKey = getAppKey(newComponentKey);
   const type = componentType.toUpperCase();
 
-  const contents: ContentUsage[] = [];
-
-
-  editorResults.forEach((result) => {
-    const currentContent = insertAndGetSummaryContent(contents, result);
-    setMultiUsage(currentContent, result);
-
-    if (currentContent.multiUsage && currentContent.multiUsage.length === 1) {
-      currentContent.hasMultiUsage = false;
-      if (!currentContent.error && currentContent.multiUsage[0].error) {
-        currentContent.error = currentContent.multiUsage[0].error;
-      }
-      currentContent.multiUsage = [];
-    }
-  });
+  const contents = buildContentResult(editorResults);
 
   const model = {
     title: `${PAGE_TITLE} - REPLACEMENT SUMMARY: ${taskSummary}`,
@@ -411,18 +352,3 @@ export function post(req: XP.Request): XP.Response {
     body: render(COMPONENT_VIEW, model),
   };
 }
-
-const setHasMultiUsage = (currentContent, wantedValue: boolean) => {
-  if (currentContent.hasMultiUsage === !wantedValue) {
-    log.warning(
-      "Mix-up on a content result, tried setting .hasMultiUsage to " +
-        JSON.stringify(wantedValue) +
-        " but it has already been set to " +
-        JSON.stringify(currentContent.hasMultiUsage) +
-        ". currentContent=" +
-        JSON.stringify(currentContent),
-    );
-    throw Error("Parameter error");
-  }
-  currentContent.hasMultiUsage = wantedValue;
-};
